+++ Starter/android/app/src/main/java/com/starter/MainApplication.java
@@ -8,7 +8,8 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
-
+import com.clevertap.android.sdk.CleverTapAPI;
+import com.clevertap.android.sdk.ActivityLifecycleCallback;
 import java.util.Arrays;
 import java.util.List;
 
@@ -41,6 +42,8 @@ public class MainApplication extends Application implements ReactApplication {
 
   @Override
   public void onCreate() {
+    CleverTapAPI.setUIEditorConnectionEnabled(false);
+    ActivityLifecycleCallback.register(this); 
     super.onCreate();
     SoLoader.init(this, /* native exopackage */ false);
   }