+++ ios/CleverTapReact/CleverTapReactManager.m
@@ -6,10 +6,13 @@
 
 #import <React/RCTLog.h>
 
-#import <CleverTapSDK/CleverTap.h>
-#import <CleverTapSDK/CleverTapSyncDelegate.h>
-#import <CleverTapSDK/CleverTapInAppNotificationDelegate.h>
-
+#import "CleverTap.h"
+#import "CleverTap+Inbox.h"
+#import "CleverTap+ABTesting.h"
+#import "CleverTapEventDetail.h"
+#import "CleverTapUTMDetail.h"
+#import "CleverTapSyncDelegate.h"
+#import "CleverTapInAppNotificationDelegate.h"
 
 @interface CleverTapReactManager() <CleverTapSyncDelegate, CleverTapInAppNotificationDelegate> {
 }
@@ -34,6 +37,10 @@
         CleverTap *clevertap = [CleverTap sharedInstance];
         [clevertap setSyncDelegate:self];
         [clevertap setInAppNotificationDelegate:self];
+        [clevertap setLibrary:@"React-Native"];
+        [clevertap registerExperimentsUpdatedBlock:^{
+            [self postNotificationWithName:kCleverTapExperimentsDidUpdate andBody:nil];
+        }];
     }
     return self;
 }